Create SunflowerWrap component and replace images
