+++ src/App.jsx
@@ -1,20 +1,12 @@
 import './generated/tailwind.css'
 import Heading from './components/Heading'
-import sunflower from './img/sunflower.png'
-import MediaQuery from 'react-responsive'
 
 const hashtagStyles = {
   sizes: 'xl:text-5xl lg:text-5xl md:text-4xl sm:text-3xl text-lg',
   margins: 'xl:mt-24 lg:mt-8 md:mt-12 sm:mt-8 mt-6'
 }
 
-const sunflowerStyles = {
-  image: 'object-contain',
-  sizes: 'xl:h-88 lg:h-72 md:h-64 sm:h-48 h-32',
-  margins: 'xl:-mt-20 lg:mt-16 md:mt-16 sm:mt-16 mt-12',
-  leftSpecific: 'xl:-mr-32 transform -rotate-45',
-  rightSpecific: 'xl:-ml-32 transform scale-x-flip xl:rotate-45',
-}
+import SunflowerWrap from './components/SunflowerWrap'
 
 function App() {
   const appClass = `h-full w-full bg-cream
@@ -31,23 +23,11 @@ function App() {
         <p>Come back then to RSVP and receive all wedding details!</p>
       </div>
 
-      <div className='flex xl:flex-row flex-col xl:justify-around xl:my-24'>
-        <MediaQuery minDeviceWidth={1280}>
-          <img
-            src={sunflower}
-            alt="left sunflower graphic"
-            className={`${sunflowerStyles.image} ${sunflowerStyles.sizes} ${sunflowerStyles.margins} ${sunflowerStyles.leftSpecific}`}
-          />
-        </MediaQuery>
+      <SunflowerWrap className='xl:my-24'>
 
         <h2 className={`${hashtagStyles.sizes} ${hashtagStyles.margins}`}>#2haveand2hoopes</h2>
 
-        <img
-          src={sunflower}
-          alt="right sunflower graphic"
-          className={`${sunflowerStyles.image} ${sunflowerStyles.sizes} ${sunflowerStyles.margins} ${sunflowerStyles.rightSpecific}`}
-        />
-      </div>
+      </SunflowerWrap>
     </div>
   );
 }
